Jarvis/requirements.txt
# Dependencias básicas
SpeechRecognition==3.10.0
pyautogui==0.9.54
requests==2.31.0
psutil==5.9.5

# Reconocimiento de voz
openai-whisper>=20231117

# Síntesis de voz
torch>=2.0.0
torchaudio>=2.0.0
TTS>=0.13.0
pygame>=2.0.0

# Procesamiento de audio
numpy>=1.21.0
scipy>=1.7.0

# PyAudio (especial para Windows)
PyAudio==0.2.11
